// src/app/services/api/model/get-book-dto.ts
/**
 * EfCoreLibraryAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GetBookDto { 
    id?: number;
    title?: string | null;
    releaseYear?: number | null;
    isbn?: string | null;
    authorId?: number;
    authorName?: string | null;
    authorFirstName?: string | null;
}